KVM: x86 emulator: expand decode flags to 64 bits (bnc#821070).
suse-commit: 787de07716b93ae19dfd2d20b3db2de1b9d26d1b

+++ arch/x86/include/asm/kvm_emulate.h
@@ -246,7 +246,7 @@ struct decode_cache {
 	struct operand dst;
 	bool has_seg_override;
 	u8 seg_override;
-	unsigned int d;
+	u64 d;
 	int (*execute)(struct x86_emulate_ctxt *ctxt);
 	int (*check_perm)(struct x86_emulate_ctxt *ctxt);
 	unsigned long regs[NR_VCPU_REGS];
